%[BugId: 160]%
Add --with-compiletools option in lofar_qatools to invoke LOFAR/autoconf_share/compiletool as the compiler. This script logs the compile command and invokes the command.

The profile variant uses the --with-compiletools options.

The compiletool script currently only logs the compile line for binaries and libraries
(via the LIBTOOL variable). It does not log invokations of CC or CXX (which are used to
produce object files).