Revised Request as follows:
* cache_override renamed to service_flag
* calc_deriv renamed to eval_mode, with a new mode called DISCOVER_SPIDS (-1)

When discover_spids is set, Node::discoverSpids() is called instead of
Node::getResult(). Default action is to collect spids from children into
a spid map; MeqParm revised to return an initial spid map for its funklets.

Service_flag now overrides cache behaviour. A request with its service flag
set bypasses the normal masked comparisons, and is only served from cache
if the request id is exactly equal (i.e., caching is only done for multiple
parents.)

Finally, the Solution symdep has been renamed into Sequence. This symdep is
generated by ReqSeq. There are no nodes that explicitly depend on it, but
it is useful because it ensures that sequenced rqids are unequal, and
thus service flags are handled correctly in the case of, e.g., multiple
solvers.