BugID: 597
When a test program fails with an exit code other than 3 (test skipped), assay
copies the output of this program to a .err file (so that the cause of the
failure can be investigated). This file must be cleaned up, however, when a
test program is known to fail (XFAIL_TESTS) in a distribution, otherwise `make
distcheck' will complain.